Select a location from the map has been handled

+++ app/src/main/java/com/giraffe/weatherforecasapplication/model/repo/Repo.kt
@@ -17,7 +17,7 @@ class Repo private constructor(private val remoteSource: RemoteSource,private va
         }
     }
 
-    override suspend fun getForecast(lat: Int, lon: Int) = remoteSource.getForecast(lat, lon)
+    override suspend fun getForecast(lat: Double, lon: Double) = remoteSource.getForecast(lat, lon)
     override suspend fun insertForecast(forecast: ForecastModel) = localSource.insertForecast(forecast)
     override suspend fun deleteAllForecasts() = localSource.deleteAllForecasts()
