+++ app/src/main/java/com/giraffe/weatherforecasapplication/model/repo/RepoInterface.kt
@@ -6,7 +6,7 @@ import retrofit2.Response
 interface RepoInterface {
 
 
-    suspend fun getForecast(lat:Int,lon:Int): Response<ForecastModel>
+    suspend fun getForecast(lat:Double,lon:Double): Response<ForecastModel>
     suspend fun insertForecast(forecast:ForecastModel):Long
     suspend fun deleteAllForecasts()
 }
