+++ app/src/main/java/com/giraffe/weatherforecasapplication/network/RemoteSource.kt
@@ -4,5 +4,5 @@ import com.giraffe.weatherforecasapplication.model.ForecastModel
 import retrofit2.Response
 
 interface RemoteSource {
-    suspend fun getForecast(lat:Int,lon:Int): Response<ForecastModel>
+    suspend fun getForecast(lat:Double,lon:Double): Response<ForecastModel>
 }
